output_connectors/gcp_cloud_storage/requirements.txt
google-cloud-storage==1.23.0
ubiops-connector==1.0.2
